src/requirements.txt
altair==5.5.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.22.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
seaborn==0.13.2
selenium==4.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
streamlit==1.41.1
sympy==1.13.3
tenacity==9.0.0
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.61
ultralytics-thop==2.0.13
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
